requirements.txt
affine==2.4.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
dash==3.3.0
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fiona==1.10.1
Flask==3.1.2
fonttools==4.61.0
GeoAlchemy2==0.18.1
geopandas==1.1.1
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
laspy==2.6.1
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
narwhals==2.12.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.4
open3d==0.19.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rasterio==1.4.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.2
shiboken6==6.10.0
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.3
wcwidth==0.2.14
Werkzeug==3.1.4
widgetsnbextension==4.0.15
zipp==3.23.0
